Task: Refactor tests to match new endpoints - Create a test class for the Account Controller and move account specific tests from the user controller test class - Refactor endpoint of the tests to the new registration endpoint - Add mockbeans to the userservice test class
